优化若干
1、取消了Script Pics路径配置，直接使用一个YGOPro路径来定位Script和Pics的路径。
2、尝试增加新建卡片、删除卡片功能。增加卡片功能还缺一个卡密输入对话框。
3、为CardList的增删改三个操作后增加一个保存数据库和刷新列表操作以解决列表数据同步刷新问题。
4、下个版本考虑去掉保存数据功能，已经没有必要了。

+++ YGOProDevelop/Model/datas.cs
@@ -60,14 +60,14 @@ namespace YGOProDevelop.Model
             }
         }
 
-        public short Level {
+        public long Level {
             get {
                 byte[] bytes = BitConverter.GetBytes(level ?? 0);
                 short lv = BitConverter.ToInt16(bytes, 0);
                 return lv;
             }
             set {
-                byte[] bytes = BitConverter.GetBytes(level ?? 0);
+                byte[] bytes = BitConverter.GetBytes(level??0);
                 byte[] vbs = BitConverter.GetBytes(value);
                 bytes[0] = vbs[0];
                 bytes[1] = vbs[1];
